Implement add light window

+++ src/light.rs
@@ -1,8 +1,9 @@
+use serde::{Deserialize, Serialize};
 use vector3d::Vector3d;
 
-#[derive(Default)]
+#[derive(Default, Deserialize, Serialize)]
 pub struct Light {
-    pub coordinates: Vector3d<isize>,
+    pub coordinates: Vector3d<f64>,
     pub minimum_beam: u16,
     pub maximum_beam: u16,
     pub name: String,
